whisper-api/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
python-multipart==0.0.9
openai-whisper==20231117
